Progress on peer dependencies
So what i'm doing here is mainly changing all libraries that should be peer dependencies but are yet still required in order for the kit to work when ran itself into peer and moving the actual dependencies to the root package. This I believe is the correct approach from now aside from having these packages in an example pacakge which ends up pulling in the deps when running against the kit itself.