Configurable Stories (#61)
* Initial work for configurable stories

- Seperated out from storybook cleanup
- Changes the default import of stories to a proxy function that handles the filtering

* Change configurable stories to use a namedExport and object config

Now instead of the default import being changed to a function, the function
can be imported as a named function like:

```
import { configureStories } from '@liquid-state/ui-kit';
```

Also included is changing the config to an object instead of just using more
paramaters on a function. The object currently supports the keys:

- name (was called prefix in the code)
- storiesOf (to be removed)
- filters

* Progress on peer dependencies

So what i'm doing here is mainly changing all libraries that should be peer dependencies but are yet still required in order for the kit to work when ran itself into peer and moving the actual dependencies to the root package. This I believe is the correct approach from now aside from having these packages in an example pacakge which ends up pulling in the deps when running against the kit itself.

* Experimental reworked default export to work without having to pass storiesOf instance